+++ composeApp/src/commonMain/kotlin/fr/vetbrain/vetnutri_mp/ViewModel/SettingsViewModel.kt
@@ -36,8 +36,8 @@ class SettingsViewModel(
                 fr.vetbrain.vetnutri_mp.Repository.ConsultationRepository? =
                 null
 ) {
-    private val _uiScale = MutableStateFlow(1f)
-    val uiScale: StateFlow<Float> = _uiScale.asStateFlow()
+    private val _uiScale = MutableStateFlow(1.0)
+    val uiScale: StateFlow<Double> = _uiScale.asStateFlow()
 
     // Résultat de l'importation
     private val _importResult = MutableStateFlow<ImportResult?>(null)
@@ -47,8 +47,8 @@ class SettingsViewModel(
     private val _isApiImporting = MutableStateFlow(false)
     val isApiImporting: StateFlow<Boolean> = _isApiImporting.asStateFlow()
 
-    private val _apiImportProgress = MutableStateFlow(0f)
-    val apiImportProgress: StateFlow<Float> = _apiImportProgress.asStateFlow()
+    private val _apiImportProgress = MutableStateFlow(0.0)
+    val apiImportProgress: StateFlow<Double> = _apiImportProgress.asStateFlow()
 
     private val _apiImportLogs = MutableStateFlow<List<String>>(emptyList())
     val apiImportLogs: StateFlow<List<String>> = _apiImportLogs.asStateFlow()
@@ -115,20 +115,20 @@ class SettingsViewModel(
     }
 
     /** Définit l'échelle de l'interface utilisateur */
-    fun setUiScale(scale: Float) {
-        val newScale = scale.coerceIn(0.5f, 2f)
+    fun setUiScale(scale: Double) {
+        val newScale = scale.coerceIn(0.5, 2.0)
         _uiScale.value = newScale
         AppSizes.adjustSize(newScale)
     }
 
     /** Augmente l'échelle de l'interface utilisateur */
     fun incrementUiScale() {
-        setUiScale(_uiScale.value + 0.1f)
+        setUiScale(_uiScale.value + 0.1)
     }
 
     /** Diminue l'échelle de l'interface utilisateur */
     fun decrementUiScale() {
-        setUiScale(_uiScale.value - 0.1f)
+        setUiScale(_uiScale.value - 0.1)
     }
 
     /** Ouvre le menu latéral */
@@ -197,12 +197,12 @@ class SettingsViewModel(
     // Contrats de suivi pour import API
     fun startApiImport() {
         _isApiImporting.value = true
-        _apiImportProgress.value = 0f
+        _apiImportProgress.value = 0.0
         _apiImportLogs.value = emptyList()
     }
 
-    fun updateApiImportProgress(progress: Float) {
-        _apiImportProgress.value = progress.coerceIn(0f, 1f)
+    fun updateApiImportProgress(progress: Double) {
+        _apiImportProgress.value = progress.coerceIn(0.0, 1.0)
     }
 
     fun appendApiImportLog(message: String) {
